requirements.txt
google-cloud-monitoring
pandas
tabulate
matplotlib
numpy
pydantic